Fixes #132 - Added a clone operation
This clones the AR, attaches clones of the
originals host_parameters and then renders the "new" action, ready for editing.

The host_parameters must not use validate_presence of :host_id when using nested attributes.
Created a hidden :nested attribute to the form to flag this behaviour

Also converted the host _form to use a table so that error boxes around the attributes look correct.
Unmodified attributes in the clone are also highlighted ready for editing

Ohad has committed ontop of the original patch, but the it still under
Paul name so he could be blamed for all :)