Fixes #540 - Manages sub proxies
Automates the creation of DHCP, DNS and TFTP settings on remote servers.
This handles all required settings in order to prepare a machine for provisioning.

DHCP:
* Reservation will be created per host, setting the following options:
* Hostname, ip, mac, boot server, boot filename

If an admin interface exists (service processor, ilom etc) a reservation can be created for it as well.

DNS:
* Primary record and reverse
* SP / iLOM record

TFTP:
* Fetch required boot files (kernel, initrd)
* Setup syslinux(pxelinux) config file based on the host MAC address.
NOTE: fetching of boot files are currently limited to http(s)/ftp media.

This patch introduce the following items:
* Generic Queue and Task in order to schedule smart proxies orchestration
* handle rollbacks in case of a proxy failure
* handle sub systems in respect to the object state ( new host, update host, delete host etc).
* Auto detection of host state, which means that hosts which don't have a proxy configure
(via its subnet or domain) will not be triggering any changes.