fixes #685 - records per page should be configurable
This patch adds a new setting how many records should be per page.

note that since the code is in the ActiveRecord::Base class, and that
its in lib (e.g. loaded once) you would need to restart foreman everytime you change this value.

we could switch to require_depenecy or other solutions if required.