fixes #1151 Fix dashboard pie, to contain correct total number of hosts
to me this makes sense. Don't shoot me if the logic isn't 100% correct...

in short, active hosts are counted as ok hosts, which makes the
counters get higher than the number of total hosts sometimes.
To bring the pie to the sum of the hosts, only the missing reports
should be added, which is included too.

Tim

Signed-off-by: Tim Speetjens <tim.speetjens@gmail.com>