fixes #402 -  Add noops/pending to metric handling, and provide pending hosts for dashboard views.
This change adds the ability to see noops in metric reports inside
foreman graphs, charts and lists. The business reason for this is to allow
users to run their puppet agents in noop mode, and see where nodes have
converged within the foreman dashboard and various foreman reports.

The status data was always stored in the database, we just had to modify
the views and model to get at it.

This change modifies the dashboard adding the new state 'pending' so it appears
in the pie graph, and adds pending to the report view in the dashboard as well.