Ensure a User with hostgroup restrictions cannot View or Edit other Hostgroups.
Ties in with my earlier patch to alter permissions - a User must
have edit_classes to change the classes in a hostgroup, and
must have create/edit/destroy_parameters to alter that part of
the Hostgroup page. Edit Hostgroups suffices for changing things like
the name and the proxy.

Fixes #1484 and Refs #1324