fix for undefined constant virt error when running
When unattended is set to false, foreman is unable to run correctly.

This patch loads virt gem anyway (as bunder force it to be installed),
this is not ideal, but since we are migrating away from virt to use
fog to manage libvirt, this should have little effect long term.