fixes #1595: show success on test connection
Added css class "btn_success" to the test connection button when the connection
is successful.

This still leaves behind a problem with ec2 (and probably rhev and vmware) where
testing the connection resets the selected region (or, datacenter for rhev and
vmware).

I'll file a separate issue in redmine for that.