fixes #1842 - Getting hostgroups via the API takes an eternity.
This changes a bit the API resonspe as well, as it makes little sense in a rest
context to return full nested objects (when the child objects can be very large).

therefore, instead of environments and puppetclasses (and eventually smart vars)
the environment and classes are now reference by id.