foreman-config - CLI tool for setting Foreman configuration
Script for listing all settings and getting or setting a value for a key.

See `foreman-config -h` for details.

The main use-case for this script is in configuration time (e.g. in Puppet
manifest setting Foreman). It uses the model layer (instead of Rest API)
because some settings (such as oauth) need to be set up before API is possible
to use in some scenarios.

Loading Rails environment is quite an expensive operation. This change allows
to set more values with one command e.g.:

    script/foreman-config -k oauth_active -v true\
                          -k oauth_consumer_key -v key\
                          -k oauth_consumer_secret -v secret