bundler_ext initial support
This adds support for bundler_ext. The idea behind this is we want to avoid
Bundler in katello foreman builds, therefore in our RPMs we are gonna rename
Gemfile to Gemfile.in.

If Gemfile.in is found, bundler is not used and loading of dependencies is
handed over to bundler_ext which loads only system gems. More info on the
bundler_ext github page.

Please note this patch does not change Foreman default behavior. Also it was
agreed Aeolus team will rename bundler_ext's namespace - in the next version
it is gonna be BundlerExt only (wihtout Aeolus stuff), becuse now this
project is used in three other projects. I will prepare a change once this
is done.