Fixes 1978 adding oVirt authentication source
This patch adds an oVirt authentication source, to be used by the
oVirt-foreman UI plugin.

The login should be in the format:
ovirt_USER@DOMAIN
Where ovirt_ is a constant prefix
USER is the oVirt user
DOMAIN is the oVirt domain

The oVirt user name indeed consist of both the username and the domain,
so basically one needs to pass ovirt_USERNAME
(USERNAME=user@domain)

The password is an authentication token, which should be a valid REST
token (REST Session ID).