Ignore value of User.current when authorizing
One exception is test environment, where the User.current is used to
simulate user being logged in.

In other environments, it's a security issue, in combination of not
clearing the thread values, it's possible to use the API without
specifying any credentials, provided the same thread handled some
other user before the attackers request.