Turn off the apipie validations for now
The main reason for the validations to be implemented in Apipie was
the ability to be able to check the documentation against the code:
i.e. it can discover inconsistencies between the documentation and
the actual code that uses the API.

However, it seems, that for now there are some weaknesses in the
implementation. Also the duplication with the ActiveRecord validations
is not helping.

Therefore let's turn the validation off by default and have the issues
fixed outside of production environment.