fixes #2023 - Don't try to save os=nil in the fact importer
I've tried to address all possible places where empty set of facts (e.g. we
dont have the facts just yet as its a brand new install) could lead to the
operating system value disappering.

i believe the issue happened because of a boolean == string compareing (e.g.
some left overs in cache might lead to if "false" which is true.

I've also ensured that we get an exception if the operating system value is empty
or if it does not pass validations.