refs #1991 - Cache not cleared prevent location / org feature to show up
1. ensures settings cache is removed when app starts
we can't remove all cache, with Rails.cache.clear as that might lead to issues
with people using memcache with multiple foreman instances (e.g. we can clear
the progress bar cache while restarting another foreman instance).

The other alternative is to put a TTL on the settings, but I think this
option for this specific case makes more sense.

2. fixed the cache directory, from /tmp to /tmp/cache.