fixes #2277 - adds spice html5 support
This also adds a newer version of the web socket proxy (where we
probably need to use an external package instead of vendor it).

we still allow users to use the old XPI based spice client (which would
ensure end to end security as well) via a 'new window' button.
that window would only be allowed currently if you are on firefox on
linux.

Paired with: Amos Benari <abenari@redhat.com>