fixes #2269 - Add gettext framework to foreman.
This patch adds a gettext based localization framework to the Foreman. This is done by adding the fast_gettext and gettext_i18n_rails gems along with their depdencies. An example translation is done to the login page and the users controller.

Next steps, once this patch is merged, is to begin to internaitonalize new pages and pages which are touched during enhancements. Then, the strings can be exported to a translation tool such as transifex or zanata.