fixes #2944 - nic type selection for libvirt
defaults to 'virtio'

I needed this in my own setup because virtio was being buggy and would
crash my apt mirror VM. switching to e1000 fixed.

I don't particularly care for the device models being hardcoded. I would
of course prefer that it be queried from the libvirt itself, but I don't
see a way to do that.