add selectable network device model for libvirt
defaults to 'virtio'

I needed this in my own setup because virtio was being buggy and would
crash my apt mirror VM. switching to e1000 fixed.

I don't particularly care for the device models being hardcoded. I would
of course prefer that it be queried from the libvirt itself, but I don't
see a way to do that.

So, the list I compiled is the list from my personal system. Maybe this
should be a Setting somewhere?

The list was generated via the commands fonud here:
http://libvirt.org/formatdomain.html#elementsNICSModel

I ran it on an ubuntu 13.04 Craring) machine