fixes #3669 - migration: minimize amount of User model exposed to migration
This commit contains a minor change to an existing migration to
minimize the amount of the User model that is exposed to the
migration to only include what is 'required' by the migration.

The purpose of this change is to ensure that we can include
other plugins (e.g. katello engine) and be able to run
db:migrate without hitting conflicts.  Conflicts include
business logic used by the model which is reliant on schema
that will be introduced by later migrations.