Fixes 3790: Adding functionality to support isolate_namespace engines.
When engines are defined with isolate_namespace, their routes and
route resolutions are not directly available to the parent application.
Thus, this adds the ability to specify a path directly by the plugin
for a given menu item. Further, since isolated engines may include
entire menus of their own and wish to separate the items, a divider
function is added.