Update apipie-rails version
The 0.0.24 version includes a fix for DOS vulnerability. Neither the
Foreman nor Katello are affected because they use Apipie cache
in the production environment.

However, when talking about the dynamic documentation for foreman
plugins, there might be a need to turn off the cache (although,
I would still prefer to regenerate the cache after a new plugin is
added than generating the docs on demand, as we do in development.)

I just wanted to make sure we don't forget about this.