Fix for border-collapse:collapse tables in Firefox
This little fix addresses the problem with border-collapse:collapse,
where some lines dissapear on tables. Cellspacing would also fix this
but it's deprecated and styles outside of CSS.
There is a bug open upstream for this:
https://bugzilla.mozilla.org/show_bug.cgi?id=688556

I tried setting the style for .table > tbody > tr:after and before,
but it did not fix the problem.

http://projects.theforeman.org/issues/4120