Fixes #4226: Provides a Rake task for plugins that compiles their assets and provides access to those assets within the core Rails asset pipeline.
This commit aims to provide three things.

1. Plugin Asset Compilation - using work done in the Katello plugin,
a task is provided that works generally for any plugin as long as the
plugin defines their assets according to the documentation.

2. Adding Asset Digests - configuration is provided in production
that reads the plugins provided asset manifest file and adds those
digests to Rails master list of digests.

3. Serving Static Assets - If Rails is configured to serve static assets,
which is usually done for testing production mode locally, plugin assets
will also be served up.