fixes #8465 - fix 500 error with invalid credentials
The client.regions assignment doesn't actually *test* anything.

Once you do client.regions.count (or .inspect or whatever) it actually makes
the connection and gets the 401, so it fails miserably.

Previously the first place it was being accessed was the form where it was
loading the regions for the select_f. Now its in the test_connection().

It occurs to me that perhaps if an API key is invalidated *after* it has been
setup and working, it will cause problems, but I think that is something that
several of these models will fall victim to.

~tommy