Pass model type so search_for is called on Host
Basically the same as 8c0888af061e459c57ae768c49dad0f83196e70b

At least on version 1.6.1, the absence of this second parameter leads to a
runtime crash when it's being verified via if the current user is allowed to
add a parameter to an unmanaged host via the APIv2. Apparently, this is only
triggerable when the user has roles which filters have search restrictions.

The root cause of the crash is basically that search_for is called on
Host::Base by app/controllers/base_controller.rb:find_nested_object

Not sure this is necessary in newer versions, up to the developers to figure
it out :)