fixes #8739 - use global setting when hostgroup root_pass is blank
A blank (not nil) root_pass on hostgroups was preventing correct inheritance
from settings in both hosts and nested hostgroups.

root_pass was also being copied from parents/settings to hostgroups when they
were saved, when it should remain blank to continue inheriting.  Only hosts
are meant to copy the password on save.