Fixes #8786 - Show override button after removing overridden parameter
After a parameter is overridden in a host or host group, if one wants to
revert the decision and return to the inherited parameter, it was
necessary to reload the page after clicking on remove. This commit
fixes that by making it show the override button again without reloading.

(cherry picked from commit 90b00443a6fe27366d6a17572b8b0a41b4dcab97)