Fixes #8919 - extract the model-specific resource names into the models
This allows plugins also to specify non-default resource name, and
prevent things like STI to mess with the resource type.

In foreman-tasks, the task has STI and although there is
ForemanTasks::DynflowTask, in permissions we don't want to have it
separated, similarly as we don't want different permissions for all
types of operating systems.