Fixes #4629 - Migrate UI to Patternfly
Patternfly is a project that enhances UIs by providing a common design
and UX, with visual and behavioral patterns based on UX research. It
stems from Bootstrap 3, so we can adapt it to Foreman with relative
ease.

Some of the changes this commit include:
  - font-awesome, Patternfly icons and OpenSans font
  - Remove spinner gifs and substitute them for CSS animations
  - Move to a 1400px wide layout, as recommended by a Patternfly
  contributor on the foreman-dev mailing list
  - jNotify alerts now use Patternflied styles.
  - Empty pages (welcome.html.erb) use Patternfly empty state pattern

I'm probably missing a lot of changes here, please feel free to suggest
more on the pull request on Github