fixes #9775 - always load Encryptable when key's missing, log runtime warning
a59972c3 causes Encryptable to be loaded before the encryption_key.rb
initialiser and the majority of the class was skipped as the key was undefined.

Now Encryptable always loads, but logs at runtime if the key is unavailable,
allowing it to be defined a bit later.

(cherry picked from commit 1fcea0e919384f9f0f384d450ecac571d5953c82)