fixes #10020 - add next-server/filename attrs to provision NIC DHCP record
A host's DHCP record used to be constructed with PXE attributes in
Orchestration::DHCP, but moved to Nic::Managed during 43c4bd7.  Nic::Managed
didn't add PXE attributes, so this commit adds these when provision? is set.

Nic::Bootable could add PXE attributes but was unused, so it is now deprecated.
PXE bootable interfaces should be of type Nic::Managed with provision: true.