fixes #9873 - generate unique alert mails for each user group member
To create distinct mails, new Mailer instances are required instead of using
the same one - else, the last message changes the previous ones.  The recipient
list is now determined in the ReportImporter, and the MailNotification helps
create Mailers for each recipient.

The locale is also set correctly for each recipient this way, ensuring their
subject lines are now also translated.