Fixes #10230 - Host clone uses create_hosts permission
Non-admin users are not able to clone hosts. This is a regression, in
the past the create_hosts permission has been used for this, but it no
longer works. Checking the controller route only fixes the regression.

(cherry picked from commit 3f13099ed9437a800a1a793f8a75c7397807994a)