Fixes #10487: Add custom logging support.
This change moves away from the Rails logger to the 'logging' gem
to support more configurability of logging. Further, this allows for
the creation of custom loggers that plugins can then register through
settings.

For example, the plugin katello could define within its settings file
(settings.plugins.d/katello.yaml) the following:

```
:katello:
  :loggers:
    :pulp_rest:
      :enabled: false
    :cp_rest:
      :enabled: true
```

Which would define 2 loggers, pulp_rest and cp_rest where the latter
is turned on be default, but the former can be enabled such that
the log file isn't continuously filled with messages from that particular
layer.