fixes #10507 - wait for completion of update env request
When testing the multiple host environment update, the integration test
submits the modal popup form and then the test ends.  The server thread
takes some time to update all host envs, add audit entries etc, but in
the meantime, the test thread begins to clean up fixtures and so can hit
a deadlock with the server thread which is still processing.