Fixes #11025: Allow skip test runner to run through Jenkins test task.
The Jenkins CI gem implements its own custom test runner. When trying
to use a custom test runner in the jenkins environment, this test
runner needs to be inherited from and included as the last argument.
In order to do this, a new task that adds our test runner is executed
before each test task so that the custom test runner is always the last
argument and wins over the base Jenkins runner.

The interface for users calling this test tasks should remain the same,
only the output (and ability to run the custom test runner in Jenkins)
should change.