Refs #3917 - Feed strong_params attributes with Apipie
This pull request enables strong_params for only Puppetclass which has
very few dependencies. The idea is to use what is introduced here in
application_shared and slowly keep on adding more models to
strong_params.

'safe_params' simply takes the documentation from Apipie that is written
in the Api#v2 controllers and uses it to generate an array that Rails
'permit' will accept. This allows us to add new controllers to
strong_params just using the API documentation