Fixes #11056: Configure base SETTINGS for test environment.
This change introduces the idea of having a base configuration for
SETTINGS within the test environment. This is captured in the
settings.yaml.test file. To be able to detect the Rails environment
in settings.rb, the ordering had to be changed whereby the Gemfile
is loaded up via boot.rb and then Rails is required so that the
Rails environment is available by the time settings.rb is processed.