Fixes #11200: Add plugin test tasks for local and Jenkins.
This adds a set of plugin test tasks that can be utilized by any
plugin. The first task is so that any number of available plugin's
tests can be run singularly or in combination alongside the Foreman
core test suite. The second task does the same but for use on Jenkins.
The last set of tasks generates a test task automatically for every
plugin currently known for running just that plugins test suite.

The rake tasks assume that plugins will have a simple test/ directory,
but if a plugin needs some customization they can specify their test
task entry point via the plugin registration.