fixes #11996 - use preload to avoid multiple fact/host table joins
When retrieving fact_values as a non-admin user, the my_facts scope
performs a join to the hosts table.  The .includes(:host) also performs
a join via eager loading, and when both are combined with a scoped
search on host.hostgroup (joined via hosts to hostgroups), the three
joins with table renaming cause errors.