Fixes #12068 - Prepend puppetclass or variable to lookup_key_id
Parameters for some objects such as EnvironmentClass, are passed as
simply lookup_key_id. On Rails 3 parameters are not as strongly checked
and it works just fine.

On Rails 4 parameters will depend on the associations of the model.
This means we should specify what kind of lookup_key is it.