Fixes #12004, #11999 - Remove reserved words from Environment friendly_id
Restores use of friendly_id, but removes 'new', 'edit' etc from the
reserved words list to ensure they can be used as identifiers.

Reverts the id-name parameterization to purely name, which fixes
related bugs that caused the parameter value to be used as a Puppet
environment name in class imports, rather than the resource name.